Make a semi functional tic tac toe game
Did a few things with the factory functions. Started first making a Row
factory function, and then changed to a Line factory function. That way
I can use it for rows, columns, and diagonals. Did a factory function
also for the mark space. That way the same object can be used in a row,
column and diagonal line without having to change the string value of it
three times.